\name{TERGMdirTrans-package}
\alias{TERGMdirTrans-package}
\alias{TERGMdirTrans}
\docType{package}
\title{
\packageTitle{TERGMdirTrans}
}
\description{
\packageDescription{TERGMdirTrans}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TERGMdirTrans}
\packageIndices{TERGMdirTrans}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{TERGMdirTrans}

Maintainer: \packageMaintainer{TERGMdirTrans}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
